Line refactored. Tested
