+++ projects/atft/src/lib/actor/data-center/server/server-stand-actor.component.ts
@@ -20,10 +20,11 @@ import { AbstractServerActor } from './abstract-server-actor';
       </atft-empty>
 
       <atft-text-mesh [text]="label" [size]="2" [bevelEnabled]="false" height="0" [centered]="true"
-                      material="basic" materialColor="0xDADADA" [translateY]="-11" [translateZ]="0.2">
+                      material="basic" materialColor="0xDADADA" [translateY]="-11" [translateZ]="0.1"
+      [depthWrite]="false">
       </atft-text-mesh>
-      <atft-frame-mesh *ngIf="showFrame" [thickness]="1" [sizeX]="15" [sizeY]="15" [translateZ]="0.1" material="basic"
-                       [materialColor]="color">
+      <atft-frame-mesh *ngIf="showFrame" [thickness]="1" [sizeX]="15" [sizeY]="15" [translateZ]="0" material="basic"
+                       [depthWrite]="false" [materialColor]="color">
       </atft-frame-mesh>
     </atft-empty>
   `
