+++ projects/atft/src/lib/object/text/text-mesh.component.ts
@@ -80,7 +80,7 @@ export class TextMeshComponent extends AbstractLazyObject3D {
   receiveShadow = true;
 
   @Input()
-  depthWrite = true;
+  depthWrite = false;
 
   @Input()
   centered = true;
